Labs/Lab_2/flask_app/requirements.txt
flask>=3.0.0
psycopg2>=2.9.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
